chore(icon-base): replace prepublish hook with manual script
As this repository is developed in npm@5, the 'prepublish' hook was being
used to perform PRE-PUBLISH tasks (e.g. lint, test, transpile). However, it will
still be executed via 'npm install' on older versions of NPM by a consumer of
the package. This will cause the package installation to fail, as development
dependencies/tasks are not intended to be downloaded/executed on 'install'.

There is a lot of confusion around this:
https://github.com/npm/npm/issues/16685
https://github.com/npm/npm/issues/15147

At the time of writing, there is no lifecycle hook available to perform a task
ONLY prior to publishing, that is backwards compatible.

A manual 'validate' script has been created to be used in it's place.

Any publishing activity will need to make sure this script is run across all
packages before the 'lerna publish' script is run. It should also be used in
place of 'lerna bootstrap'.